Install a Bing Maps API Key in Your '.Renviron' File for Repeated Use

This function will add your CENSUS API key to your .Renviron file so it can be called securely without being stored in your code. After you have installed your key, it can be called any time with get_bing_maps_api_key().

Usage

bing_maps_api_key(
  key,
  overwrite = FALSE,
  install = FALSE,
  default = "BING_MAPS_API_KEY"
)

get_bing_maps_api_key(
  key = NULL,
  default = "BING_MAPS_API_KEY",
  call = caller_env()
)

Arguments

key

The API key provided to you from the Bing Maps Dev Center formatted in quotes. A key can be acquired at https://www.bingmapsportal.com/

overwrite

If this is set to TRUE, it will overwrite an existing BING_MAPS_API_KEY that you already have in your .Renviron file.

install

if TRUE, will install the key in your .Renviron file for use in future sessions. Defaults to FALSE